+++ main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"os"
@@ -25,13 +24,9 @@ func main() {
 		}
 		return
 	}
+	var debug bool
 	if len(os.Args) > 1 && os.Args[1] == "debug" {
-		err := plugin.Debug(context.Background(), "registry.terraform.io/databrickslabs/databricks",
-			&plugin.ServeOpts{ProviderFunc: provider.DatabricksProvider})
-		if err != nil {
-			log.Fatal(err.Error())
-		}
-		return
+		debug = true
 	}
 	log.Printf(`Databricks Terraform Provider (experimental)
 
@@ -40,5 +35,8 @@ Version %s
 https://registry.terraform.io/providers/databrickslabs/databricks/latest/docs
 
 `, common.Version())
-	plugin.Serve(&plugin.ServeOpts{ProviderFunc: provider.DatabricksProvider})
+	plugin.Serve(&plugin.ServeOpts{
+		ProviderFunc: provider.DatabricksProvider,
+		Debug: debug,
+	})
 }